# docs/swagger.yaml
basePath: /
definitions:
  main.Activity:
    description: 配對活動資訊
    properties:
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/main.Location'
      location_id:
        type: integer
      target_count:
        type: integer
      title:
        type: string
    type: object
...
  main.Location:
    description: 地點資訊
    properties:
      address:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 這是一個買一送一配對網站的API文檔
  title: 買一送一配對網站 API
  version: "1.0"
paths:
  /admin/activities:
    get:
      consumes:
      - application/json
      description: 取得所有配對活動的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Activity'
            type: array
        "500":
          description: 無法取得活動列表
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 取得配對活動列表
      tags:
      - 管理員
    post:
      consumes:
      - application/json
      description: 建立新的配對活動
      parameters:
      - description: 配對活動資訊
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/main.Activity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Activity'
        "400":
          description: 無效的請求資料
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法建立活動
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 建立新的配對活動
      tags:
      - 管理員
  /admin/activities/{id}:
    delete:
      consumes:
      - application/json
      description: 刪除指定ID的配對活動
      parameters:
      - description: 活動ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 活動已刪除
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 無效的活動 ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法刪除活動
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 刪除配對活動
      tags:
      - 管理員
    put:
      consumes:
      - application/json
      description: 更新指定ID的配對活動
      parameters:
      - description: 活動ID
        in: path
        name: id
        required: true
        type: integer
      - description: 配對活動資訊
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/main.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Activity'
        "400":
          description: 無效的請求資料
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法更新活動
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新配對活動
      tags:
      - 管理員
  /admin/locations:
    get:
      consumes:
      - application/json
      description: 取得所有地點的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Location'
            type: array
        "500":
          description: 無法取得地點列表
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 取得地點列表
      tags:
      - 管理員
    post:
      consumes:
      - application/json
      description: 建立新的地點
      parameters:
      - description: 地點資訊
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/main.Location'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Location'
        "400":
          description: 無效的請求資料
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法建立地點
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 建立新的地點
      tags:
      - 管理員
  /admin/locations/{id}:
    delete:
      consumes:
      - application/json
      description: 刪除指定ID的地點
      parameters:
      - description: 地點ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 地點已刪除
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 無效的地點 ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法刪除地點
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 刪除地點
      tags:
      - 管理員
    put:
      consumes:
      - application/json
      description: 更新指定ID的地點
      parameters:
      - description: 地點ID
        in: path
        name: id
        required: true
        type: integer
      - description: 地點資訊
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/main.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Location'
        "400":
          description: 無效的請求資料
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 無法更新地點
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新地點
      tags:
      - 管理員
...
securityDefinitions:
  ApiKeyAuth:
    description: 輸入 'Bearer &lt;JWT token&gt;' 進行認證。可以先透過 Facebook 登入取得 token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"